Refactor amounts from type int to float

+++ productionCalculatorLib/components/nodes/interfaces/IHasProduct.cs
@@ -5,5 +5,5 @@ namespace productionCalculatorLib.components.nodes.interfaces;
 public interface IHasProduct
 {
     Product Product { get; set; }
-    int Amount { get; set; }
+    float Amount { get; set; }
 }
