+++ productionCalculatorLib/components/nodes/interfaces/IHasRecipe.cs
@@ -5,5 +5,5 @@ namespace productionCalculatorLib.components.nodes.interfaces;
 public interface IHasRecipe
 {
     Recipe Recipe { get; set; }
-    int ProductionAmount { get; set; }
+    float ProductionAmount { get; set; }
 }
